fix: handle delete own user

+++ server/src/features/users/delete-user/delete-user-controller.ts
@@ -8,6 +8,14 @@ export const DeleteUserController = {
   deleteUser: async (req: Request, res: Response) => {
     const id = req.params.id;
 
+    const currentUserId = req.user?._id.toString();
+
+    if (id === currentUserId) {
+      return res
+        .status(400)
+        .json(ApiResponse.failed('Admin cannot delete own account'));
+    }
+
     const result = await DeleteUserService.deleteUser(id);
 
     res
